Remove RestUtils
